Pull ToggleShowButton into PasswordEntry
